Generate category HTML from search select on Create/Update campaign page
This is just UI elements only. These will be checked when the form is submitted to get the JSON data to submit.
- New table to show the selected categories
- Function to generate the html for each category row along with it's depth control and remove button
- Select2 search box tag mode removed, just a single selection
- Same changes made to both Create and Update campaign templates
Note: The old input has now been hidden, but is populated when adding new categories. This is just to keep the system working while we make the transition to JSON data in a new hidden field.

Bug: T224325
Change-Id: I48c724850a78547fe9268fbb6c7e554034a8a68e