Add click event to expand/collapse metadata
Very simple solution all wrapped in the event handler will is good enough for current phase.
As the button text is being changed on expand/collapse, we have both states stored in JavaScript.
Not ideal as we need to same text exactly to be served in the HTML.
Best to fix once multi lingual UI is working using either:
- switch to CSS :after for button text based on the "expand" classes we add/remove
- switch to JS only for the expand/collapse button text

Bug: T226111
Change-Id: Ic08b88b055ab1d92069cd03ca4fc8324b6a1f949