Setup basic webpack bundling for front end JavaScript
Use "npm i" in the console to install dependencies, then "npm run dev" to build the files.
Bundles are generated for each page specified in our entry list (webpack.config file).
The bundles are added to the static/js folder, with the same name as their source js file.
The old static JS files have been left in place for this commit but will be replaced when you run the build command.
No functionality has been changed, all parts of the app should be working as before.
- Add package.json and webpack.config files
- Setup "npm run dev" command to watch for changes
- Setup "npm run build" to minify files and optimise for production (i.e. what we deploy to live)
- All JavaScript source files moved to new src folder and converted to modules
- ParticipationManager, wiki-loves-countries and utils split off as separate files
- Add static/js to gitignore as the output files are now generated by webpack

Bug: T226339
Change-Id: I87b1142f3d88e772c169d9bc27571f55d95aeb2c