Store existing statement IDs
We need to know the unique id for each statement we display to the user in the UI.
This is used for "remove" and "mark prominent" actions, as in both cases we need to alter the data about an existing statement.
Not applicable for newly added statements as they only get an id assigned once the edit is complete on the wiki.
The following should be stored once the Structured Data API call is complete for each image:
- Statement IDs for each existing depict statement, so it can be sent to server as part of the contribution data
- Media ID for the image, also sent to server with each contribution
- Revision ID of the page (note, this is probably better to retrieve when getting edit tokens but storing now just in case we need it)
  Note: The revision ID will change after each individual contribution is saved

Bug: T226092
Change-Id: Id42ccf4459aa595dda3342b13a56c846458f6867