Collect images on the server side when a campaign is updated
This replaces the client side API requests when a user visits
`/campaigns/???/participate`. New images are fetched when a campaign
is created and each time it is updated, even if no changes are made to
the categories. The requests are done in separate threads and in
serial rather than parallel (which was the case in the previous
solution).

If the campaign is for a Wiki Loves contest the countries are also
added to the database. Each images in the campaign is assigned a
country.

If all the images for a campaign have not been loaded the user gets a
notice when trying to participate.

When the images are fetched they are retrieved from
`/campaigns/???/images` by the client. In Wiki Loves campaigns the
`/<country name>` is appended to get only images for the category for
that country.

Bug: T226303
Change-Id: I47d653599a98fa8b6ba12ad1a0c72891e6fc3275